users/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pillow==10.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose==3.3.0
python-multipart==0.0.6
pyjwt==2.8.0
pandas==2.1.3
requests==2.31.0
uvicorn[standard]
pillow==10.2.0
aiohttp==3.10.0